Add line drawing algorithm and implement hough
* Implement Bresenham's Line drawing algorithm.
* Implement hough detector that is more efficient that the traditional
  3D version.

The x and y gradients developed during edge detection are used to
determine the direction of the line perpendicular to the edge at
each point. The line is extended and values are given to each
x,y coordinate the line drawer goes through drawing it.

The values are normalized and the maxima will be extracted. These
will be the circle centers. After the centers are found, the radii
need to be discovered.